+++ public/js/ui/premissas.js
@@ -2,7 +2,7 @@ import { P } from '../nucleo/estado.js';
 import { $, num } from '../nucleo/formato.js';
 
 const CAMPOS = ["dens","tch","plantio","arr_ha","hdia","disp","dias","diasOper","diasTrab","hTurno",
-"diesel","arr","adm","imob","dep","ipreco","densCarga","volTransb","velC","velV","tCarga",
+"diesel","arr","imob","dep","ipreco","densCarga","volTransb","velC","velV","tCarga",
 "tDesc","hDiaTr","dispTr","consTr","manutTr","raioSafra","raioMuda","perdaCarga","desnivel","rendBomba",
 "kwh","fonte","hPorMec","eqPorAjud","colPorLider","tercAereaTar","tercSistTar","tercSistHa","tercOutros"];